fix(ci): python-full-coverage — split oversized ownership-edges tests under 1500-line gate
